requirements.txt
python-telegram-bot==21.8
aiosqlite
pytz
python-dotenv
